Added memory_region
memory.h: header file with memory_region and utility functions.
memory_region is meant to serve as a single place in memory that stores
all types of objects the raytracer uses.
You basically make objects (spheres, triangles, in the future I will add
materials and texures), and call a function to copy them to the
memory_region. Then you get back a pointer to that object.
I think this will be useful because it lets you resuse things like
materials and textures and encapsulates the allocation and freeing.

Still teseting it.