Completely refactoring IntersectFBMSphere.
The previous implementation was incorrect because it made fbm spheres have different  surfaces fromdifferent points of view..

Added a macro definition for USE_RAYMARCH_FBM. Use the new function if defined else use the old one.

Also theres a function IntersectConfettiSphere which came about from an error in the new IntersectFBMSphere. I'll try to salvage it later because it makes interesting shapes.